requirements.txt
numpy<2.0
torch==2.1.2
torchvision==0.16.2
gymnasium==0.29.1
faiss-cpu==1.7.4
networkx==3.2
owlready2==0.46
openai
anthropic
google-generativeai
# HuggingFace stack — pinned for PyTorch 2.1.2 compatibility (transformers 5.x requires PyTorch>=2.4)
transformers==4.41.2
sentence-transformers==3.0.1
accelerate==0.30.1
huggingface-hub==0.23.4
tokenizers==0.19.1
ragas==0.1.21
datasets==2.19.0
pandas
scipy
matplotlib
seaborn
tqdm
rich
tensorboard
python-dotenv
python-docx
pypdf
jupyter
ipykernel
pytest==8.2.2
pytest-cov==5.0.0